requirements.txt
minio>=7.1.7
label-studio-ml>=1.0.8
python-dotenv>=0.19.1
redis>=4.3.4
gunicorn==20.1.0
rq==1.10.1
uvicorn>=0.18.2
fastapi>=0.78.0
pydantic>=1.8.2
requests>=2.28.1
loguru>=0.6.0
